Refactor UploadCard component into smaller components
