+++ src/components/UploadCard/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useFileProcessor'
+export * from './useHtmlReport'
